Updated licenses field and maintainer email
Per https://wiki.archlinux.org/title/PKGBUILD#license the use of
`custom:*license-name*` causes vim to highlight the field with red
background, indicating possibly an error. The included tarball does have
a license that is very similar to OFL-1.1-no-RFN therefore this is now
updated to `LicenseRef:OFL-1.1-no-RFN` which will not cause vim to
highlight that field.

No other changes to this package, and as this license format slightly
differs from the one listed on SPDX, it is still included in the
subsequent package, as "OFL".

Maintainer email changed in this commit.