Set `RUSTUP_TOOLCHAIN` in `prepare`, too
When running `cargo pkgid` in `prepare`, set `RUSTUP_TOOLCHAIN` to fix
the following error that occurs if the user has multiple toolchains
installed:

```
==> Starting pkgver()...
error: rustup could not choose a version of cargo to run, because one wasn't specified explicitly, and no default is configured.
help: run 'rustup default stable' to download the latest stable release of Rust and set it as your default toolchain.
```

Suggested-by: Coelacanthus