Point to 0.6.x branch instead
The 'master' branch is EOL, and will be replaced by the 'avalonia'
branch at a later point.

The upstream supported git version is on the 0.6.x branch, which this
commit targets instead.

The plan is to have this package point to the active development branch,
so once 'avalonia' development continues, this package may end up
pointing to that instead, and this package will be spun off, though
nothing is set in stone yet