Fix patch set for latest libxml2 release.
The patch itself is suboptimal, it technically isn't bad but because a
non-function always-present version symbol was added, a null check is a dead
code now. Maybe I'll update it, maybe don't. ManaPlus is dead already tbh
and now what this does it to prolong the inevitable: moving towards ManaVerse
fork for people. who are somehow stuck with the ManaPlus.