Add fix for cuda 13. Disable libnpp.
npp filters (like scale_npp, from libnpp) are not compatible with
cuda 13.

A comment[1] from a ffmpeg developer indicates that libnpp is to
be considered deprecated and that npp filters will not receive any
compatibility fixes. He also said that libnpp eventually may be
disabled in future ffmpeg commits by the configure script if it
detects a newer libnpp version.

[1] https://code.ffmpeg.org/FFmpeg/FFmpeg/issues/20191#issuecomment-6514