Fix use of --no-as-needed in LDFLAGS
Previously --no-as-needed was appended to LDFLAGS assuming the last
argument was a -Wl,... and that it was OK to extend it.
This is a bit sketchy and now it seems to have broken the build
(probably due to some side change in Arch/makepkg/make/ld/...).

Append an entire `-Wl,--no-as-needed` to fix the build.