Work around build error
GCC has recently started to fail building this package due to the following
error. A pull request to fix this issue has been proposed upstream (see
link) but has not yet been merged. While we wait for an upstream solution
to arrive, we can apply the same fix with a bit of sed magic.

Link: https://github.com/DIGImend/hidrd/pull/33/commits/73a01e5e0091f5fbf452fd85ac3632f8c0171e65