refactor PGP signature check
the package has been relying on a historical feature of --verify, where
gpg will assume that the first argument is a detached signature and will
strip certain suffixes to try to find the signed data.

gpg discourages using this historical feature and recommends specifying
the signed data directly. this updates the PGP check to use gpg's
--verify as intended, with two arguments and adds AgileBits's PGP key
fingerprint to the 'validpgpkeys' list (this package doesn't
specify a .sig file as a source, so this is just for informational
purposes only)