Initial commit
I had to do a lot of hacking to get the hwinfo module to build correctly
because the provided makefile is bad. If the repo changes the way hwinfo
builds it might break this PKGBUILD in the future.

Since I got that module working you should be able to see the fan speeds
and temperature just by running: sensors.

Also provides two scripts:

  - su_axb35_monitor
  - su_axb32_info

See also: https://strixhalo-homelab.d7.wtf/Guides/Power-Mode-and-Fan-Control