Initial upload: termshot 0.6.0-1
This clobbers a much older project by the same name, but that has been
deleted from the AUR for many years and hasn't been touched upstream
for seven. I think this Go based project has a better claim to the
package name, so I'm going ahead with posting this on top of the old
namespace.