Add optional dependency on ffmpeg for hx decoding support
ndi-sdk handles decoding of HX codecs (H.264 and H.265) through
ffmpeg[1].

The ffmpeg libraries are not linked, but they are hard coded in
libndi.so (with version suffix), which means that probably they
are loaded by dlopen().

Current ffmpeg library versions hard coded indicate ffmpeg 7.1
(libavcodec.so.61 and libavutil.so.59), which can be retrieved
by strings|grep piped commands.

References
----------
[1] https://docs.ndi.video/all/developer-guides/decoding-with-ndi