wireguard-vanity-keygen: update in regard of guidelines
Create the package with the current Go packaging guidelines. This silences some namcap warnings.

Additionally remove the extra source url as it doesn't really benefit the cause.
Remove $srcdir as this is the default starting location.
Adjust arch as this isn't an any package. This contains a compiled binary tied to your architecture.
Add the suffix to the renamed tarball.