Bump 2.4.2 + fix for device_discovery.cc
Upon bumping to version 2.4.2 it appears there might be WIN32 specific
code left outside of ```#IFDEF _WIN32 ... #ENDIF``` preprocessor
directives.

Adding ```device_discovery-fix.patch``` to address said issue and a
pull request has been created to have the fix applied in the upstream
here - https://github.com/dev47apps/droidcam-obs-plugin/pull/30

While it's a bit dangerous to include a patch for a -git package I'm
doing so in the interim so that this package can be on the latest 2.4.2
and successfully build.  I'll continue to monitor the pull request I
opened above on the upstream and will remove the patch when the fix
is merged.