Don't build pixbuf loader and thumbnailer
gdk-pixbuf2 upstream recommends [1], on Linux, to use glycin for image
loading and to also rely on glycin for thumbnailing. gdk-pixbuf2 on Arch
Linux depends on glycin, which now includes a loader and thumbnailer for
JXL files which works. However Arch Linux still builds libjxl with its
pixbuf loader and thumbnailer.

This change follows upstream's recommendation and disables building
libjxl's pixbuf loader and thumbnailer.

[1] https://gitlab.gnome.org/GNOME/gdk-pixbuf/-/releases/2.43.2#changes-since-242