Disable the mime plugin
After commit 3d27ac68845, only the mime plugin was left,
as the gimp and gdk-pixbuf plugins are now disabled.

But according to upstream, the mime plugin is not needed
anymore, as shared-mime-info 2.2 supports the `image/jxl`
mime type.[1][2]

So, now all plugins can be disabled at once with
`JPEGXL_ENABLE_PLUGINS`, and there is no more need
for individual plugin options in cmake.

[1] https://github.com/libjxl/libjxl/blob/v0.11.1/plugins/mime/README.md
[2] https://github.com/libjxl/libjxl/commit/b931fc50ddf296e4c0908c8883c6c5dc53031c61