upgpkg: 0.0.3-15: only use CPM for necessary packages
Unti now, CPM was being used to download everything it possibly can.
After reading the CPM docs, it was decided to disable CPM, and have it
look for local packages first. It still uses CPM for stuff like boost,
spirv-tools, and cubeb. In the end, the build is now shorter a few
hundred files