Explicitly set scale to work around missing `GetScreenScale` function.
A recent PathOfBuilding commit[1] started making use of the
SimpleGraphic[2] API which isn't available in the Linux port. As a
work-around the call can be removed an replaced with a literal `1.0`
screen scale.

I'm not exactly sure if the API is missing from `pobfrontend`[3] or
`PathOfBuilding-runtime-src`. Once added, this patch should be reverted
to allow for screen scaling.

[1] https://github.com/PathOfBuildingCommunity/PathOfBuilding/commit/e7f33688431a5db47a7e432db146e322f7c379e7#diff-81ab5df67490eff4be50a40b5501d36f6a6b8ea59cb3b8943a08f98de0f776a7
[2] https://github.com/PathOfBuildingCommunity/PathOfBuilding-SimpleGraphic
[3] https://github.com/ernstp/pobfrontend.git#branch=qt6
[4] https://github.com/Openarl/PathOfBuilding/files/1167199/PathOfBuilding-runtime-src.zip