Refactor and update to 1.1.9.9
Rehosting was probably against the terms of the license, and the download is behind a login. So just make the user do
that themselves. Refactor the install process to mirror what the provided install script would have done had we used
that. Utilize the provided .desktop file and launcher program instead of a homebrewed one. Become compliant with newer
packaging guidelines.