exempt gnome-shell from FakeRoot
Certain versions of the gnome-shell binary refuse to run as root, only
outputting an error message. Not even the --version option is parsed.

As the package function runs under FakeRoot, it looks exactly like that
to the program. FakeRoot works by modifying the LD_LIBRARY_PATH and
LD_PRELOAD. By clearing the library to preload, usual function is
restored.

This can be reasonably be considered safe, as the program is only run
with a parameter to output its version number and exit, not making any
modification or communications to or with the rest of the system.