Source from extensions.gnome.org
After version 1.14, the Makefile lost the LOCALINSTALL feature. [1] The
default Make target now installs the extensions for the running user.
Alternatively, it can be packed up for review on extensions.gnome.org .

The release versioning scheme on GitHub also changed to mirror the
version number used on extensions.gnome.org at this point. Although
there are some versions missing: 1.14 is equal to 57, but the next
release on GitHub is 68. Also the latest release on extensions.gnome.org
is 83, the highest number on GitHub is 77.

So as this now seems to be the main distribution pathway and this is a
release version package (in contrast to the -git variant [2]), it makes
sense to pull releases from there.

[1]: https://github.com/dsheeler/CoverflowAltTab/pull/187
[2]: https://aur.archlinux.org/packages/gnome-shell-extension-coverflow-alt-tab-git