upgpkg: 590.48.01-1
This is thought to prepare for the update of the 590 driver, as well as replacing the "closed source" kernel module.

NVIDIA does still provide the "closed source" kernel module, but generally there seem not be much reason to still package this. The closed source module would make it possible to disable the "GSP" Firmware. According to NVIDIA the closed source modules receive way less testing and therefore it is not recommended.

Migration Strategy:

- nvidia-open-dkms replaces nvidia-dkms
- nvidia-open replaces nvidia
- nvidia-open-lts replaces nvidia-lts

The "*-open" naming scheme will be kept to avoid confusing for a bit of time. In a few year(s) as soon the closed source module gets dropped from NVIDIA's run file, it would be possible to rename it again.

If we really want to package the "closed source" module, I would suggest a new naming as "nvidia-closed-dkms" or "nvidia-legacy-dkms", as well there should be no modules built for the "linux" and "linux-lts" kernel to reduce the workload.

Users with Pascal (10xx or lower) will need to migrate manually the the 580xx branch. This PKGBUILD will be available in the AUR.
We need to write an announcement that these GPUs are **not supported** anymore. This will still do few interruptions at users not reading the NEWS, but without maintaining the "nvidia-580xx" driver in the official repository there will be no smooth migration possible.