fix ghdl source checksum
The 5.1.1 tag for ghdl seems to have moved. Specifically, the following
commits are no longer present, as the tag seems to have been moved back
by 2 commits. Specifically, these are missing:
- 8fb617d744a974dcc03970500aeaf15315b07b64
- 91725e47fdded6a3ac2e4e5ee5fa1adb4b8b4f6f

The diff is minor (NEWS.md and GitHub CI), so only updating the checksum
should be fine.

The whole thing is weird; but not much else I can do on my end.