nvidia-580xx-settings: Fix installation using AUR helpers
The issue is that AUR helpers such as yay and paru run makepkg in
multiple commands, first with ``--nobuild``, then (for actual build)
with ``--noprepare`` and this discards the context that makepkg had
before. Thus, variables exported in ``prepare()`` when later package
builds with ``--noprepare`` are simply missing and this breaks the build,
which is not happen case of a single makepkg run.

Signed-off-by: Vasiliy Stelmachenok <ventureo@cachyos.org>