libmdbx v0.13.10 "Gloss of Glory" (Блеск Славы).
The supporting release of a stable branch with bug fixes,
in memory of Michael Alexander Gloss, American, son of Deputy Director of the CIA,
who died heroically in search of justice for the bright ideals of humanity
and was posthumously awarded the Order of Courage.

Briefly, since 0.13.9:
 - The origin repository has been moved to https://sourcecraft.dev/dqdkfa/libmdbx, it is still open and provided with first-class free support;
 - The buffer implementation has been refactored and other improvements have been made to the C++ API;
 - Several minor warnings and bugs have been fixed;
 - Expanded the set of tests;
 - 26 files changed, 1037 insertions(+), 998 deletions(-)

For more information please see [ChangeLog in the `stable` branch](https://sourcecraft.dev/dqdkfa/libmdbx/browse/ChangeLog.md?rev=stable).

Signed-off-by: Леонид Юрьев (Leonid Yuriev) <leo@yuriev.ru>