Build extension
After a series of restructuring commits [1], the extensions needs to be
composed before it can be installed. The build step is the place to do
so.

Fortunately, a Make directive doing this was added, too. Unfortunately,
it spits out the extension as an archive. So it is unpacked for the
package process to find it.

[1]: https://github.com/dsheeler/CoverflowAltTab/pull/187